sber-code-docs-mcp/requirements.txt
mcp
llama-index-core
llama-index-readers-file
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss
sentence-transformers
faiss-cpu
